flask/requirements.txt
click==8.1.3
click-log==0.4.0
colorama==0.4.6
Flask==2.2.3
Flask-JWT-Extended==4.5.2
gitdb==4.0.10
GitPython==3.1.31
httpcore==0.16.3
httpx==0.23.3
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==23.13.1
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
pkginfo==1.9.6
postgrest==0.10.6
psycopg2-binary==2.9.7
PyJWT==2.8.0
typing_extensions==4.5.0
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3